requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-dateutil
